Move playback_test (now gone) to a testutil
